- fix-compiler-warnings

+++ libeplayer3-arm/output/writer/mipsel/vp.c
@@ -136,7 +136,7 @@ static int writeData(WriterAVCallData_t *call, bool is_vp6, bool is_vp9)
 		UpdatePesHeaderPayloadSize(PesHeader, payload_len);
 		// it looks like for VUPLUS drivers PES header must be written separately
 		int ret = call->WriteV(call->fd, iov, 1);
-		if (iov[0].iov_len != ret)
+		if (iov[0].iov_len != (unsigned)ret)
 			return ret;
 		ret = call->WriteV(call->fd, iov + 1, 1);
 		return iov[0].iov_len + ret;
@@ -150,12 +150,12 @@ static int writeData(WriterAVCallData_t *call, bool is_vp6, bool is_vp9)
 		int bytes = payload_len - 10 - 8;
 		UpdatePesHeaderPayloadSize(PesHeader, payload_len);
 		// pes header
-		if (pes_header_len != WriteExt(call->WriteV, call->fd, PesHeader, pes_header_len)) return -1;
+		if (pes_header_len != (unsigned)WriteExt(call->WriteV, call->fd, PesHeader, pes_header_len)) return -1;
 		if (bytes != WriteExt(call->WriteV, call->fd, call->data, bytes)) return -1;
 
 		offs += bytes;
 
-		while (bytes < call->len)
+		while ((unsigned)bytes < call->len)
 		{
 			int left = call->len - bytes;
 			int wr = 0x8000;
@@ -177,7 +177,7 @@ static int writeData(WriterAVCallData_t *call, bool is_vp6, bool is_vp9)
 
 			UpdatePesHeaderPayloadSize(PesHeader, wr + 3);
 
-			if (pes_header_len != WriteExt(call->WriteV, call->fd, PesHeader, pes_header_len)) return -1;
+			if (pes_header_len != (unsigned)WriteExt(call->WriteV, call->fd, PesHeader, pes_header_len)) return -1;
 			if (wr != WriteExt(call->WriteV, call->fd, call->data + offs, wr)) return -1;
 
 			bytes += wr;
@@ -209,7 +209,7 @@ static int writeData(WriterAVCallData_t *call, bool is_vp6, bool is_vp9)
 		PesHeader[29] = 0xFF;
 		PesHeader[33] = 0x85;
 
-		if (pes_header_len != WriteExt(call->WriteV, call->fd, PesHeader, 184)) return -1;
+		if (pes_header_len != (unsigned)WriteExt(call->WriteV, call->fd, PesHeader, 184)) return -1;
 
 		return 1;
 	}
